+++ client/src/components/Header.tsx
@@ -3,8 +3,8 @@ import { useDispatch, useSelector } from 'react-redux'
 import { LinkContainer } from 'react-router-bootstrap'
 import { Navbar, Nav, Container, NavDropdown } from 'react-bootstrap'
 
-import { AppState } from '../../types'
-import { logout } from '../../redux/actions/user'
+import { AppState } from '../types'
+import { logout } from '../redux/actions/user'
 
 const Header = () => {
   const dispatch = useDispatch()
