+++ client/src/components/Message.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Alert } from 'react-bootstrap'
 
-import {MessageProps} from '../../types'
+import {MessageProps} from '../types'
 
 const Message = ({ variant, children }: MessageProps) => {
   return <Alert variant={variant}>{children}</Alert>
